Explicit mention of parameter value for `unite_imu_method`
There is confusion regarding what value needs to be passed to `unite_imu_method`. The `ros2-development` branch expects a different value for the `unite_imu_method` parameter as compared to the `ros2-legacy` and `ros1-legacy` branches. I have added this detail in the README. 

Thanks to @Nir-Az for pointing this out in [this](https://github.com/IntelRealSense/realsense-ros/issues/2542#issuecomment-1318831114) issue.